requirements.txt
tensorflow==2.15.0
streamlit
numpy
pillow
